pyproject.toml
[tool.poetry]
name = "podcast-guest-research"
version = "0.1.0"
description = "A web interface for automating podcast guest research"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
typing-extensions = "^4.5.0"
requests = "^2.31.0"
langgraph = "^0.0.2"
langchain-core = "^0.0.10"
langchain-openai = "^0.0.2"
langchain-community = "^0.0.2"
langchain-google-genai = "^0.0.1"
tiktoken = "^0.4.0"
aiohttp = "^3.8.5"
apify-client = "^1.0.0"
google-api-python-client = "^2.100.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
tavily-python = "^0.2.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 